node

+++ controllers/contactController.js
@@ -50,6 +50,11 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error("contact not found");
   }
 
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
+
   const updateContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -67,6 +72,12 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("contact not found");
   }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
+
   await Contact.deleteOne();
   res.status(200).json(contact);
 });
